requirements.txt
Django==1.11.4
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
twilio>=3.3.6
django-twilio==0.8.0
django-storages==1.1.8
boto==2.38.0
django-import-export==0.4.5
